// legacy/components/table-filter/types.ts
export interface FieldBaseItem {
  /** 元属性 */
  readonly attributes?: Readonly<any>;
  key: string;
  /** 字段标签 */
  label?: string;
  /** 表单中不展示label */
  hideLabel?: boolean;
  /** @deprecated 已移除，采用栅格方案 */
  minWidth?: number | string;
  /** @deprecated 已移除，采用栅格方案 */
  maxWidth?: number | string;
  /** Group相同的表单项在交互上合并成一个，比如 按下单日期/按收货日期 */
  group?: string;
  /** 默认为false，为true在“全部筛选条件”中初始时为勾选状态 */
  defaultUsed?: boolean;
  /** 默认为false，为true不允许在“全部筛选条件”中取消勾选 */
  alwaysUsed?: boolean;
  /** 是否隐藏 */
  hide?: boolean;
  placeholder?: string;
  /** 是否禁用 */
  disabled?: boolean;
  /** Antd, 展示清空的icon按钮，默认为true */
  allowClear?: boolean;
  /** 是否折叠 */
  collapsed?: boolean
  /** 自定义组件 */
  render?: React.ReactNode;
}

type API_Method = (r?: any, codes?: number[]) => Promise<any>;
type ToParam<Option, API extends API_Method> = (
  value: Option,
  pre: Partial<Exclude<Parameters<API>[0], undefined>>,
) => Partial<Exclude<Parameters<API>[0], undefined>>;

/** 表单项-输入框 */
export interface FieldInputItem<API extends API_Method = API_Method> extends FieldBaseItem {
  type: 'input';
  /** Antd， 输入框类型 */
  inputType?: 'number' | 'text';
  /** 默认值 */
  defaultValue?: string;
  /**
   * 自定义如何merge本字段的键值到params中，默认merge { [FieldItem.key]: value }到params中
   *
   * 如果value为undefined（字段没有输入），不会进入此函数，而是直接忽略此字段
   */
  toParam?: ToParam<FieldInputItem<API>['defaultValue'], API>;
}
// ...
export type SelectOptions = Array<{
  text: string;
  value: number | string;
  /** 分组名，相同展示为一组 */
  group?: string;
}>;

export interface FieldSelectItem<
  API extends API_Method = API_Method,
  Multiple extends boolean = boolean,
  S extends SelectOptions = SelectOptions,
> extends FieldBaseItem {
  type: 'select';
  multiple?: Multiple;
  allowClear?: boolean;
  /**
   * 触发搜索的方式，默认为`manual`；
   *
   * - `manual` 点击“查询”按钮后才查询；
   * - `onChange` 字段表单变化后查询，且查询和重置按钮会被隐藏；
   * - `both` 字段表单变化后查询，点击“查询”按钮后查询；
   * - `onBlur` 失去焦点时查询，可以使用 onBlur 事件触发查询；
   */
  trigger?: 'onChange' | 'manual' | 'both' | 'onBlur';
  /**
   * 列表选项，支持数组、函数、异步函数
   *
   * 支持远程搜索
   *
   * Gm_api的List接口支持直接用helper.ts中的{@link generateListOptions}方法转换
   */
  options?: S | ((searchVal?: string) => S) | ((searchVal?: string) => Promise<S>);
  /** 默认值，不传为空，为空提交时忽略此字段 */
  defaultValue?: Multiple extends true ? S[0]['value'][] : S[0]['value'];
  /** 远程搜索模式，默认为false即前端筛选模式 */
  remote?: boolean;
  // 限制多选时的最大选择数量
  maxLength?: number;
  selectProps?: SelectProps;
  /** 同 {@link FieldInputItem.toParam} */
  toParam?: ToParam<Multiple extends true ? S[0]['value'][] : S[0]['value'], API>;
  /** 是否在第一次的挂载的时候进行获取option */
  isMountToFetch?: boolean;
}

/** 表单项-日期选择器 */
export interface FieldDateTypeItem<API extends API_Method = API_Method> extends FieldBaseItem {
  type: 'date';
  /** 区间型选择器 */
  range?: false;
  defaultValue?: Moment;
  /** 同 {@link FieldDateRangeItem.showTime} */
  showTime?: PickerDateProps<Moment>['showTime'];
  /** 同 {@link FieldDateRangeItem.disabledDate} */
  disabledDate?: PickerDateProps<Moment>['disabledDate'];
  /** Antd, 设置选择器类型 */
  picker?: PickerMode;
  /** 同 {@link FieldInputItem.toParam} */
  toParam?: ToParam<Moment, API>;
  allowClear?: boolean;
}

/** 表单类型，日期范围选择器 */
export interface FieldDateRangeItem<API extends API_Method = API_Method> extends FieldBaseItem {
  type: 'date';
  /** 区间型选择器 */
  range: true;
  defaultValue?: RangeValue<Moment>;
  /** Antd,预设时间范围快捷选择 */
  ranges?: RangePickerProps['ranges'];
  /** Antd,增加时间选择功能 */
  showTime?: RangePickerDateProps<Moment>['showTime'];
  /** Antd,不可选择的日期 */
  disabledDate?: (date: Moment, { begin, end }?: { begin?: Moment; end?: Moment }) => any;
  /** Antd, 设置选择器类型 */
  picker?: PickerMode;
  /** 打开选择器时是否清空已选择的值 */
  openClearValues?: boolean;
  /** 同 {@link FieldInputItem.toParam]} */
  toParam?: ToParam<RangeValue<Moment>, API>;
  onCalendarChange?: (val: RangeValue<Moment>, dateStrings: string[], info: any) => void;
  allowClear?: boolean;
}

export type FieldDateItem = FieldDateTypeItem | FieldDateRangeItem;

export interface CasCaderOption {
  value: string | number;
  label?: React.ReactNode;
  disabled?: boolean;
  children?: CasCaderOption[];
}

/** 表单项-级联选择器 */
export interface FieldCascaderItem<
  API extends API_Method = API_Method,
  Multiple extends boolean = boolean,
> extends FieldBaseItem {
  type: 'cascader';
  /** Antd, 多选 */
  multiple: Multiple;
  showSearch?: boolean;
  options?: CasCaderOption[] | (() => CasCaderOption[]) | (() => Promise<CasCaderOption[]>);
  defaultValue?: string[];
  /** Antd, 根结点是否可选，默认为true */
  changeOnSelect?: boolean;
  /** 同 {@link FieldInputItem.toParam} */
  toParam?: ToParam<Multiple extends true ? string[][] : string[], API>;
  showCheckedStrategy?: CascaderProps<any>['showCheckedStrategy']
  /** 使用ant 原始渲染 */
  useAntdDisplayRender?: boolean
  displayRender?: CascaderProps<CasCaderOption>['displayRender'] | null
}

/** 自定义表单项 */
export interface FieldCustomizeItem<API extends API_Method = API_Method> extends FieldBaseItem {
  type: 'customize';
  defaultValue?: any;
  toParam?: ToParam<any, API>;
  render: React.ReactNode
}

/**
 * 字段配置，使用到toParams时建议传入api,类型更安全：
 *
 *     const FIELDS: FieldItem<typeof ListSaleOutStockSheet>[] = [{...}]
 */
export type FieldItem<API extends API_Method = API_Method> =
  | FieldInputItem<API>
  // | FieldInputNumberItem<API>
  | FieldSelectItem<API>
  | FieldSelectItem<API>
  | FieldDateTypeItem<API>
  | FieldDateRangeItem<API>
  | FieldCascaderItem<API>
  | FieldCustomizeItem<API>;
// ...
export interface CachedSetting {
  [key: string]: { visible?: boolean };
}

export type CachedFields = Array<
  {
    attributes?: any;
    visible?: boolean;
  } & Pick<FieldItem, 'key' | 'label' | 'type'>
>;